Create Information Site, Update Ingredients Table

+++ database/migrations/2023_05_12_124324_create_ingredients_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('ingredients', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('price');
+            $table->decimal('price', 7, 2);
             $table->integer('amount');
             $table->string('measure');
         });
